# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.DeleteSubscriptionRequest:
    properties:
      id:
        type: string
    type: object
  dto.Subscription:
    properties:
      created_at:
        type: string
      end_month:
        type: integer
      end_year:
        type: integer
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_month:
        type: integer
      start_year:
        type: integer
      user_id:
        type: string
    type: object
  dto.SumSubsRequest:
    properties:
      from:
        type: string
      service_name:
        type: string
      to:
        type: string
      user_id:
        type: string
    type: object
  dto.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing user subscriptions
  title: Subscription Service API
  version: "1.0"
paths:
  /create_sub:
    post:
      consumes:
      - application/json
      description: Create a new subscription for user
      parameters:
      - description: Subscription data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Subscription'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create new subscription
      tags:
      - subscriptions
  /delete_sub:
    delete:
      consumes:
      - application/json
      description: Delete subscription by ID (from JSON body)
      parameters:
      - description: Subscription ID
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete subscription
      tags:
      - subscriptions
  /get_all_subs:
    get:
      description: Returns list of all subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all subscriptions
      tags:
      - subscriptions
  /sum_subs:
    post:
      consumes:
      - application/json
      description: Returns the total amount spent on subscriptions for a given user,
        service, and period.
      parameters:
      - description: Sum subscriptions request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SumSubsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Total sum
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get total subscription sum
      tags:
      - subscriptions
  /update_sub:
    put:
      consumes:
      - application/json
      description: Update subscription fields by ID (JSON body)
      parameters:
      - description: Updated data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update subscription
      tags:
      - subscriptions
swagger: "2.0"